backend/dev-requirements.txt
--requirement requirements.txt

django-stubs
djangorestframework-stubs
flake8
mypy
pylint
pylint-django
